Move dark mode toggle to setting screen
